internet_engineer/Geogram/Geogram/messenger/views.py
from rest_framework import viewsets
from .models import *
from .serializers import *

class MessageViewSet(viewsets.ModelViewSet):
    """
    This viewset automatically provides `list` and `detail` actions.
    """
    queryset = Message.objects.all()
    serializer_class = MessageSerializer