internet_engineer/Geogram/Geogram/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
